Add ActiveDoor script and change the Door script
