+++ Assets/Scripts/Door/ActiveDoor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 172ef41d909b12346a36016f01d43bb5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
